Set opt-level=1 for dev builds
These algorithms can run incredibly slow without optimization.
For 007-two-knights(n=5000), this reduces time taken from ~10s to around ~3s.
Using opt-level=2 takes around 607ms and opt-level=3 takes 362ms.

Honestly this tradeoff may not be worth it if it hurts compile times.
It is possible that improvements in tests could make up this cost,
but we could also just use a separate test target for that.