007-two-knights: Add FixedSizeSpiral iterator
This iterates over the outermost layer of a spiral for a board size/radius.
This allows BoardSpiral type to just be a composition of this with increasing sizes.

Based on a "spiral index" which gives the relative position in the FixedSizeSpiral.
Due to the use of an index, this can also occur in reverse.

Changed BoardSize to store 0-based index rather than 1-based size.

By itself, this doesn't affect much.
However, I plan to use the knowledge of the index to skip over the bulk of the inner two loops.