007-two-knights: Away from corners, possible movements are constant
My first guess, that movements=2 away from the edges was wrong.
In fact, movemeents=0 away from the *corners*.
I started to realize my mistake writing up the justification for the optimization,
For most of the the outer spiral, the only places that have not already been handled
are those spaces on the same row/column as the knight itself.
A knight cannot attack onto the same row/column as itself,
meaning all possible movements are in spaces that have already been handled.
However, I did not finally conform that my first guess was wrong until I wrote the test for it.

After writing up the main justification,
I also refactored the naive implementation to split out the movement counting into its own function.
and more clearly justified how the comparison against the index is equivalent
to checking for already handled spaces.

Finally, I implemented the test and confirmed my suspissions that movements=2 was over-counting.
The first paragraph of this commit is really a rough draft of the new justification,
now that I am certain I was wroung and know why I was (or at least think I do).

The result of this order of doing things is that the original justification
has bad reasoning and claims that movements=2 is constant,
but the test is updated reflects the behavior of checking around the corners.

I have not actually performed the optimization yet,
just made the preparations for it.
This optimization will be extremely profitable,
meaning that when the size is large the inner most loops
will only run a small constant number of times.
This will essentially turn addition in a loop into a single multiplication.

Some Observations on my skill level:

Its funny that I learned about these sets of problems from matklad's blog post:
<https://matklad.github.io/2023/08/06/fantastic-learning-resources.html>.
He said that at "level 3" you start to understand about documenting invariants
and are able to use pure reasoning to ensure your submission is correct the first time.
I would like to consider myself at level 3,
but in most cases I have had to do multiple submissions to get to the answer.
In many problems (like this one) this is because my first attempt is too slow rather than incorrect.
I am usually pretty careless about documentating/justifiying my algorithms,
and rely just as much on unit testing for correctness as pure reasoning.
Although I guess this is not too surprising as I would not claim to be
as talented or experienced a programer as matklad is ;)