007-two-knights: Fix repeated addition optimization
This converts addition in a loop into multiplication,
as I have been trying to do for the last commits.

The issue with the last attempt was that the naive and optimized
implementations have different `should_ignore` closures.
I was not taking this into account and assuming the optimized impl
has the same comparison as the navie impl (actually its inverted).

This from about 1.2 seconds to 17 ms.
I love these sorts of optimizations :D